<BadgeSelect/> - fix word wrapping and eyes issues (#2397)
This PR fixes an issue when an item text contained more than 2 words,
the text would break.

Also, this PR fix the eyes tests. Since the DropdownLayout of the
BadgeSelect is centered, it overflowed the AutoExample container, thus
cutted off the screen in the eyes tests. This PR fixed that by centering
the BadgeSelect.

---

- [ ] Change is tested
- [ ] Change is documented
- [ ] Build is green
- [ ] Change of UX is approved by [wuwa](https://github.com/wuwa) or [milkyfruit](https://github.com/milkyfruit)

### Thanks for contributing :)